Do not reinstate old tracks in JBrowse.navigateToLocation.
We always call navigatetolocation from showRegion. And we always call
showRegion with a list of tracks to display.

showRegion will always opens the tracks passed to it as parameters. And
then, when jumping between tasks on two different reference sequences,
navigatetolocation tries to re-open tracks that were previously visible.
This causes tracks to be displayed twice when Afra jumps from a task on
one ref seq to a task on another ref seq.

Ideally, opening tracks should be idempotent.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>