Change ScratchPad semantics and store helpers in EditTrack.
Make best effort to ensure features rendered in EditTrack are in the order they
appear in the store and that ordering is retained after an edit action:
JB fetches features from a store object in chunks, of say 1250bp. If a feature
in our store starts in the first chunk, no matter what order it appears in the
store it will be rendered first.

EditTrack is now refreshed only after all transcripts have been inserted in,
removed from or replaced in the store. That is, if we were trying to insert two
transcripts, and one of them causes an error, the view won't refresh. The store
will be left in an inconsistent state though.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>