Improve view-store helpers.
All view-store helpers follow transactional semantics. That is, if an error
occurs when inserting, removing, replacing transcripts, the view will not be
refreshed, nor will the store be left in an inconsistent state.

We now have separate helpers to update store when the transcript was split and
when two transcripts were merged.

This also fixes regressions in annotation editing.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>