Update Perl subsystem to JB 1.11.5.
We forked JB 1.10.2. This commit removes Perl modules and bin/*.pl that we
don't need and upgrades the rest to JB 1.11.5 pulling in JB's Perl test suite.

To guide the upgrade process in future, we add a simple script that reports
what Perl module needs to be updated/removed/added.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>