Replace JB's sequence search box with Bootstrap modal.
Sequence search box is now replaced with feature rich modals provided by
Bootstrap. Modals retain the search params and selections which is quite
handy when making quick searches, are easy to dismiss, and go with the
overall apperance of afra.

We also add a couple of handy feature to search modal like search on
'Enter' key press, and autofocus search area.

Signed-off-by: Vivek Rai <vivekraiiitkgp@gmail.com>