Sync undo/redo states across browser tabs/windows.
And restore undo/redo state along with the gene model being curated on page
reload or on closing and subsequently opening a tab.

Additionally, we now sync to localStorage from EditTrack.changed(). So that
syncing to localStorage happens after all changes have been finalised in
one tab.

Syncing starts from the first edit action. Syncing _to_ localStorage always
overwrites existing data in localStorage. But localStorage is read only and
only if the id set in localStorage corresponds to the id of the task opened
in the current browser tab. Thus if two different tasks are open in two
different browser tab / window, they won't be synchronised with each other.
Yet, no matter the number of different tasks opened in different tabs, tabs
that have the same task open will always be synchronised. Editing more than
one task at a time will affect the persistence though. That is, edit track
can't be restored on tab reload if user is simultaneously editing two tasks.

To the above, it must be noted that a user working on two tasks at a time is
not a standard use case in Afra. A user can open another task only on the
pretext of revisiting an existing submission.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>